Update style.css
Finally AFTER 3 WHOLE HOURS, I finally figured out how to get this layout like the layout from the picture. Since I spent most of my time doing this, my next goal is to get the Typography together then I will be done! 

I learned about CSS Grid, which I was making it harder than it should but it is valuable tool to use now that I have a better understanding of it. Go me! Day 8 of 100 days of code.